Add instrumentation to the Speedy machine & optimize
Performance optimizations discussed in issue #5764

Specific changes include:

- Flags: `enableClassifyInstrumentation`, `enableLightweightStepTracing'
- Running the machine: `step()` --> `run()` (`SResultContinue` --> `SResultFinalValue`)
- No more `Ctrl` type; Speedy control is now always an expression.
- We never construct full-applied PAPs (they don't make sense as a value!)

Performance improvement is in the region: x1.2 .. x1.3

Additional details:

I changed Speedy so the machine.ctrl is always an SExpr. This avoids a virtual dispatch on every step of the machine. It also roughly halves the number of steps taken (two old steps becomes one new step)

I've changed the top level control of Speedy: from machine.step() to machine.run, with the control of stepping while the machine returns SResultContinue moved into speedy itself. (And so SResultContinue is removed in favour of SResultFinalValue.) The main advantage of this approach is that the tight while loop can be moved inside the exception handler, rather than having to wrap the handler every step.

changelog_begin
changelog_end