nuke the Bazel cache
This is a bit of a nuclear option as we cannot figure out how to repair
the cache in place. It seems (somewhat) clear that the cache is
corrupted for Windows builds, but we're out of ideas on either how to
test it to fully confirm or how to repair it.

This PR will consist of two commits: one with the Bazel cache definition
commented out (this one), and then one that reverts it. This will result
in a no-change commit when squashed on master, but should serve as an
audit trail for the accompanying Standard Change (as I consider dropping
the entire cache outside of routine maintenance).

CHANGELOG_BEGIN
CHANGELOG_END