nuke the Bazel cache (#5782)
* nuke the Bazel cache

This is a bit of a nuclear option as we cannot figure out how to repair
the cache in place. It seems (somewhat) clear that the cache is
corrupted for Windows builds, but we're out of ideas on either how to
test it to fully confirm or how to repair it.

This PR will consist of two commits: one with the Bazel cache definition
commented out (this one), and then one that reverts it. This will result
in a no-change commit when squashed on master, but should serve as an
audit trail for the accompanying Standard Change (as I consider dropping
the entire cache outside of routine maintenance).

CHANGELOG_BEGIN
CHANGELOG_END

* run Windows build with no Linux node

Normally our Windows builds only start after a Linux node has run the
`check_for_release` step. This commit bypâsses that so this PR can run
while all the other ones are stuck in the Linux queue.

* Revert "nuke the Bazel cache"

This reverts commit f7cd68e0809f62b06b61ef21e4db1d5b11ef18c3, recreating
the Bazel cache.

* populate macos cache and test Windows one

* revert all changes in this PR