Better tests for environment handling.
Have tests which will fail if we screw up the indexing for any of:
- function application args
- the stack (let bindings)
- free-variables of a closure

In particular if we mix up the direction of out indexing. i.e. 0..n-1 VS n-1..0

Try to make the three tests as selective as possible, by only failing for the one aspect they are designed to test.

changelog_begin
changelog_end